# Use fewer borrows in `ExtentInner` API