This PR tweaks the `ExtentInner` API to use fewer borrows from the caller's frame.

It's a building block for making extents independent and capable of executing their own IOs in parallel.

This is _theoretically_ slightly less efficient, because it allocates an extra `Vec` or two, but I doubt there's any real impact; I'm waiting for Madrid to come back up to get some performance numbers.